refactor visual script to editor
