+++ packages/editor/src/panels/VisualScript/container/index.tsx
@@ -29,6 +29,7 @@ import { EditorControlFunctions } from '@ir-engine/editor/src/functions/EditorCo
 import { SelectionState } from '@ir-engine/editor/src/services/SelectionServices'
 import { VisualScriptComponent } from '@ir-engine/engine'
 import { getState } from '@ir-engine/hyperflux'
+import Button from '@ir-engine/ui/src/primitives/tailwind/Button'
 import { VisualScriptState } from '@ir-engine/visual-script'
 import { isEqual } from 'lodash'
 import React from 'react'
@@ -36,7 +37,6 @@ import { useTranslation } from 'react-i18next'
 import AutoSizer from 'react-virtualized-auto-sizer'
 import { ReactFlowProvider } from 'reactflow'
 import 'reactflow/dist/style.css'
-import Button from '../../../../../primitives/tailwind/Button'
 import { Flow } from '../flow'
 import './ReactFlowStyle.css'
