+++ packages/editor/src/panels/VisualScript/sidepanel.tsx
@@ -35,14 +35,14 @@ import { useTemplateHandler } from '@ir-engine/editor/src/components/visualScrip
 import { useVariableHandler } from '@ir-engine/editor/src/components/visualScript/hooks/useVariableHandler'
 import { NodetoEnginetype } from '@ir-engine/engine'
 import { NO_PROXY, useMutableState } from '@ir-engine/hyperflux'
+import SelectInput from '@ir-engine/ui/src/components/editor/input/Select'
+import StringInput from '@ir-engine/ui/src/components/editor/input/String'
+import PaginatedList from '@ir-engine/ui/src/components/editor/layout/PaginatedList'
+import NodeEditor from '@ir-engine/ui/src/components/editor/properties/nodeEditor'
+import ParameterInput from '@ir-engine/ui/src/components/editor/properties/parameter'
+import Button from '@ir-engine/ui/src/primitives/tailwind/Button'
 import { GraphTemplate, VariableJSON, VisualScriptDomain, VisualScriptState } from '@ir-engine/visual-script'
-import Button from '../../../../../primitives/tailwind/Button'
-import SelectInput from '../../../input/Select'
-import StringInput from '../../../input/String'
-import PaginatedList from '../../../layout/PaginatedList'
-import NodeEditor from '../../../properties/nodeEditor'
-import ParameterInput from '../../../properties/parameter'
-import { Examples } from '../modals/load'
+import { Examples } from './modals/load/load'
 
 type templateHandler = ReturnType<typeof useTemplateHandler>
 type variableHandler = ReturnType<typeof useVariableHandler>
